Pull out policy styles into files matching the component
