+++ styles/modules/Policy.module.scss
@@ -14,165 +14,8 @@
  * limitations under the License.
  */
 
-@import "styles/constants";
-@import "styles/mixins";
-
-@mixin formWidth {
-  width: 100%;
-
-  @include desktopAndLarger {
-    width: 75%;
-  }
-}
-
-.form {
-  width: 100%;
-  height: 100%;
-  padding: 0.5rem;
-
-  @include tabletAndLarger {
-    padding: 1rem 2rem;
-  }
-}
-
-.policyInputsContainer {
-  margin-bottom: 1rem;
-  padding: 1rem;
-  @include formWidth;
-
-  > * {
-    margin: 0.5rem auto 0;
-  }
-
-  @include desktopAndLarger {
-    margin-right: auto;
-
-    > * {
-      margin: 1rem auto 0;
-    }
-  }
-}
-
-.policyValidationContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: flex-start;
-
-  @include desktopAndLarger {
-    align-items: center;
-    flex-direction: row;
-  }
-}
-
-.validateButton {
-  margin-bottom: 1rem;
-  align-self: flex-start;
-
-  @include desktopAndLarger {
-    margin-bottom: 0;
-    margin-right: 1rem;
-  }
-}
-
-.validationResults {
-  flex: 1;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  color: $GREEN;
-  margin: 0 0.5rem;
-
-  > p {
-    margin-left: 0.25rem;
-  }
-}
-
-.failedResultsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  flex: 1;
-  width: 95%;
-
-  @include desktopAndLarger {
-    width: 55%;
-  }
-
-  .validationResults {
-    color: $HOT_RED;
-  }
-}
-
-.documentation {
-  margin-top: 0.25rem;
-  font-size: 0.75rem;
-  color: $MEDIUM_GREY;
-  text-align: right;
-}
-
-.actionButtonsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  @include formWidth;
-
-  @include tabletAndLarger {
-    flex-direction: row;
-    justify-content: space-between;
-  }
-}
-
-.actionButtons {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  > button {
-    margin: 0.5rem auto;
-  }
-
-  @include tabletAndLarger {
-    flex-direction: row;
-    justify-content: flex-start;
-
-    > button {
-      margin: 0.5rem;
-    }
-
-    > button:first-of-type {
-      margin-left: 0;
-    }
-
-    > button:last-of-type {
-      margin-right: 0;
-    }
-  }
-}
-
-.modalActionButtons {
-  @extend .actionButtons;
-  margin: 0 auto;
-  justify-content: center;
-
-  @include tabletAndLarger {
-    justify-content: center;
-  }
-}
-
 .pageContainer {
   height: 100%;
-}
-
-.playgroundButton {
-  @include tabletAndLarger {
-    align-self: flex-end;
-  }
 }
 
 .notFound {
@@ -183,38 +26,5 @@
     font-size: 1.5rem;
     font-weight: 600;
     margin: 1rem;
-  }
-}
-
-.confirmDeleteText {
-  text-align: center;
-  margin-bottom: 1.5rem;
-}
-
-.updateMessageText {
-  &:last-of-type {
-    margin-bottom: 1.5rem;
-  }
-}
-
-.darkTheme {
-  .policyInputsContainer {
-    background-color: $DT_BACKGROUND_MEDIUM;
-  }
-
-  .validationResults {
-    color: $BRIGHT_GREEN;
-  }
-
-  .failedResultsContainer {
-    .validationResults {
-      color: $HOT_RED;
-    }
-  }
-}
-
-.lightTheme {
-  .policyInputsContainer {
-    background-color: $LT_BACKGROUND_LIGHT;
   }
 }
