+++ styles/modules/PolicyForm.module.scss
@@ -1,0 +1,157 @@
+/**
+ * Copyright 2021 The Rode Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import "styles/constants";
+@import "styles/mixins";
+
+@mixin formWidth {
+  width: 100%;
+
+  @include desktopAndLarger {
+    width: 75%;
+  }
+}
+
+.form {
+  width: 100%;
+  height: 100%;
+  padding: 0.5rem;
+
+  @include tabletAndLarger {
+    padding: 1rem 2rem;
+  }
+}
+
+.policyInputsContainer {
+  margin-bottom: 1rem;
+  padding: 1rem;
+  @include formWidth;
+
+  > * {
+    margin: 0.5rem auto 0;
+  }
+
+  @include desktopAndLarger {
+    margin-right: auto;
+
+    > * {
+      margin: 1rem auto 0;
+    }
+  }
+}
+
+.policyValidationContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: flex-start;
+
+  @include desktopAndLarger {
+    align-items: center;
+    flex-direction: row;
+  }
+}
+
+.validateButton {
+  margin-bottom: 1rem;
+  align-self: flex-start;
+
+  @include desktopAndLarger {
+    margin-bottom: 0;
+    margin-right: 1rem;
+  }
+}
+
+.documentation {
+  margin-top: 0.25rem;
+  font-size: 0.75rem;
+  color: $MEDIUM_GREY;
+  text-align: right;
+}
+
+.actionButtonsContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  @include formWidth;
+
+  @include tabletAndLarger {
+    flex-direction: row;
+    justify-content: space-between;
+  }
+}
+
+.actionButtons {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  > button {
+    margin: 0.5rem auto;
+  }
+
+  @include tabletAndLarger {
+    flex-direction: row;
+    justify-content: flex-start;
+
+    > button {
+      margin: 0.5rem;
+    }
+
+    > button:first-of-type {
+      margin-left: 0;
+    }
+
+    > button:last-of-type {
+      margin-right: 0;
+    }
+  }
+}
+
+.modalActionButtons {
+  @extend .actionButtons;
+  margin: 0 auto;
+  justify-content: center;
+
+  @include tabletAndLarger {
+    justify-content: center;
+  }
+}
+
+.confirmDeleteText {
+  text-align: center;
+  margin-bottom: 1.5rem;
+}
+
+.updateMessageText {
+  &:last-of-type {
+    margin-bottom: 1.5rem;
+  }
+}
+
+.darkTheme {
+  .policyInputsContainer {
+    background-color: $DT_BACKGROUND_MEDIUM;
+  }
+}
+
+.lightTheme {
+  .policyInputsContainer {
+    background-color: $LT_BACKGROUND_LIGHT;
+  }
+}
