+++ styles/modules/PolicyValidationResult.module.scss
@@ -1,0 +1,73 @@
+/**
+ * Copyright 2021 The Rode Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+@import "styles/constants";
+@import "styles/mixins";
+
+.validationResults {
+  flex: 1;
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+  margin: 0 0.5rem;
+
+  > p {
+    margin-left: 0.25rem;
+  }
+}
+
+.failedResultsContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+  flex: 1;
+  width: 95%;
+
+  @include desktopAndLarger {
+    width: 55%;
+  }
+
+  .validationResults {
+    color: $HOT_RED;
+  }
+}
+
+.darkTheme {
+  &.validationResults {
+    color: $BRIGHT_GREEN;
+  }
+
+  &.failedResultsContainer {
+    .validationResults {
+      color: $HOT_RED;
+    }
+  }
+}
+
+.lightTheme {
+  &.validationResults {
+    color: $GREEN;
+  }
+
+  &.failedResultsContainer {
+    .validationResults {
+      color: $HOT_RED;
+    }
+  }
+}
